DAH-1089 refactor to iso component and add loading check
