+++ app/javascript/__tests__/modules/listingDetails/ListingDetailsUnitAccordions.test.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import renderer from "react-test-renderer"
-import { ListingDetailsUnitAccordion } from "../../../modules/listingDetails/ListingDetailsUnitAccordion"
-import { unitsGrouped } from "../../data/RailsListingUnits/listing-units"
+import { ListingDetailsUnitAccordions } from "../../../modules/listingDetails/ListingDetailsUnitAccordions"
+import { openSaleListing } from "../../data/RailsSaleListing/listing-sale-open"
 
 describe("ListingDetailsUnitAccordion", () => {
   it("displays the unit accordions for a given listing", () => {
@@ -19,15 +19,7 @@ describe("ListingDetailsUnitAccordion", () => {
       }
     })
     const tree = renderer
-      .create(
-        Object.keys(unitsGrouped).map((unitType) => (
-          <ListingDetailsUnitAccordion
-            key={unitType}
-            unitType={unitType}
-            unitGroup={unitsGrouped[unitType]}
-          />
-        ))
-      )
+      .create(<ListingDetailsUnitAccordions listingId={openSaleListing.Id} />)
       .toJSON()
 
     expect(tree).toMatchSnapshot()
